Add files via upload
Added balanced accuracy metric (as well as accuracy, just to see how misleading it can be when dataset is imbalanced). Also, at the end, added a couple of cells to investigate the largest discrepancies in terms of predicted probability of a given diagnosis across multiple images of a given lesion.